Size is not grandeur, and territory does not make a nation

+++ 0x00-hello_world/6-size.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 /**
- * main - The parent function of the C program
- * Return: Gives you back the void 0 you gave it 
+ * main - Write a C program that prints the size of various types on computer
+ * Return: 0
  */
 int main(void)
 {
@@ -11,9 +11,9 @@ int main(void)
 	long long int lli;
 	float f;
 
-	printf("Size of a char: %d byte(s)\n", sizeof(char));
-	printf("Size of an int: %d byte(s)\n", sizeof(int));
-	printf("Size of a long int: %d byte(s)\n", sizeof(long int));
-	printf("Size of a float: %d bytes(s)\n", sizeof(float));
+	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(char));
+	printf("Size of an int: %d byte(s)\n", (unsigned long) sizeof(int));
+	printf("Size of a long int: %d byte(s)\n", (unsigned long)sizeof(long int));
+	printf("Size of a float: %d bytes(s)\n", (unsigned long)sizeof(float));
 	return (0);
 }
